contentful-rich-text-renderer用のoptionsをts化
